requirements.txt
Flask==0.12.2
Flask-Admin==1.5.0
Flask-BabelEx==0.9.3
Flask-Cors==3.0.3
Flask-Login==0.4.1
Flask-Mail==0.9.1
flask-marshmallow==0.8.0
Flask-Migrate==2.1.1
Flask-Permissions==0.2.3
Flask-Principal==0.4.0
Flask-RESTful==0.3.6
Flask-Security==3.0.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
inflect==0.2.5
marshmallow==2.15.0
marshmallow-sqlalchemy==0.13.2
Pillow==5.0.0
six==1.11.0
speaklater==1.3
SQLAlchemy==1.2.1
SQLAlchemy-Continuum==1.3.3
SQLAlchemy-Utils==0.32.21
unicode-slugify==0.1.3
validators==0.12.0
WTForms-Alchemy==0.16.5
WTForms-Components==0.10.3
